+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/config/RedshiftDataApiConfig.kt
@@ -5,6 +5,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import software.amazon.awssdk.regions.Region
+import software.amazon.awssdk.services.athena.AthenaClient
 import software.amazon.awssdk.services.redshiftdata.RedshiftDataClient
 import software.amazon.awssdk.services.redshiftdata.model.ExecuteStatementRequest
 import software.amazon.awssdk.services.redshiftdata.model.RedshiftDataRequest
@@ -21,6 +22,15 @@ class RedshiftDataApiConfig(
   fun redshiftDataClient(): RedshiftDataClient {
     return RedshiftDataClient.builder()
       .region(Region.EU_WEST_2)
+      .build()
+  }
+
+  @Bean
+  @ConditionalOnMissingBean(AthenaClient::class)
+  fun athenaClient(): AthenaClient {
+    val region = Region.EU_WEST_2
+    return AthenaClient.builder()
+      .region(region)
       .build()
   }