+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalprisonreportinglib/data/AthenaAndRedshiftCommonRepository.kt
@@ -1,0 +1,55 @@
+package uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data
+
+import org.apache.commons.lang3.time.StopWatch
+import org.slf4j.LoggerFactory
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionResponse
+import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.redshiftdata.StatementExecutionStatus
+
+abstract class AthenaAndRedshiftCommonRepository : RepositoryHelper() {
+
+  companion object {
+    private val log = LoggerFactory.getLogger(this::class.java)
+  }
+
+  abstract fun executeQueryAsync(
+    query: String,
+    filters: List<ConfiguredApiRepository.Filter>,
+    sortColumn: String?,
+    sortedAsc: Boolean,
+    policyEngineResult: String,
+    dynamicFilterFieldId: String? = null,
+    database: String? = null,
+    catalog: String? = null,
+  ): StatementExecutionResponse
+
+  abstract fun getStatementStatus(statementId: String): StatementExecutionStatus
+
+  fun getPaginatedExternalTableResult(
+    tableId: String,
+    selectedPage: Long,
+    pageSize: Long,
+    jdbcTemplate: NamedParameterJdbcTemplate = populateJdbcTemplate(),
+  ): List<Map<String, Any?>> {
+    val stopwatch = StopWatch.createStarted()
+    val result = jdbcTemplate
+      .queryForList(
+        "SELECT * FROM reports.$tableId limit $pageSize OFFSET ($selectedPage - 1) * $pageSize;",
+        MapSqlParameterSource(),
+      )
+      .map {
+        transformTimestampToLocalDateTime(it)
+      }
+    stopwatch.stop()
+    log.debug("Query Execution time in ms: {}", stopwatch.time)
+    return result
+  }
+
+  fun count(tableId: String, jdbcTemplate: NamedParameterJdbcTemplate = populateJdbcTemplate()): Long {
+    return jdbcTemplate.queryForList(
+      "SELECT COUNT(1) as total FROM reports.$tableId;",
+      MapSqlParameterSource(),
+    ).first()?.get("total") as Long
+  }
+}